requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
diskcache==5.6.3
fpdf==1.7.2
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pycparser==2.22
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
regex==2024.11.6
requests==2.32.3
tiktoken==0.9.0
typing_extensions==4.13.0
urllib3==2.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
diskcache==5.6.3
fpdf==1.7.2
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.8
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pycparser==2.22
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-docx==1.1.2
regex==2024.11.6
requests==2.32.3
tiktoken==0.9.0
typing_extensions==4.13.0
urllib3==2.3.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
diskcache==5.6.3
executing==2.2.0
fpdf==1.7.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
llama_cpp_python==0.3.8
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
parso==0.8.4
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tiktoken==0.9.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
urllib3==2.3.0
wcwidth==0.2.13
